Add Cart and Shop page
I added in the cart icon and created a context for it. I also created the styles and component for the cart. So when you click the cart will open but no other functionality beyond that. I also added in the hats shopping page with a local json file. This will be exported out in the future.