Fix Bug
The bug of the items not rendering was fixed. It was a mistake of not returning in the cart-item component. There was also another bug with the quantity not updating and only adding in new components. There was an uppercase 'I' instead of a lowercase one that was causing the issue.